requirements.txt
slack-bolt==1.18.0
python-dotenv==1.0.0
langchain==0.0.150
langchain-huggingface==0.0.1
